requirement.txt
google-generativeai
python-dotenv
matplotlib
seaborn